% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctm_initial.R
\name{.build_trajectory_formulas}
\alias{.build_trajectory_formulas}
\title{Build trajectory formulas from degree and knots}
\usage{
.build_trajectory_formulas(outcome, time_var, degree, knots = NULL)
}
\description{
Build trajectory formulas from degree and knots
}
\keyword{internal}
