% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctm_plot.R
\name{.create_mean_plot}
\alias{.create_mean_plot}
\title{Create Mean Trajectory Plot}
\usage{
.create_mean_plot(model, time_range, n_points, ci, colors, class_labels = NULL)
}
\description{
Create Mean Trajectory Plot
}
\keyword{internal}
