% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctm_initial.R
\name{.create_residual_guide}
\alias{.create_residual_guide}
\title{Create reference guide plot with example residual patterns}
\usage{
.create_residual_guide()
}
\description{
Create reference guide plot with example residual patterns
}
\keyword{internal}
