% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctm_plot.R
\name{.create_spaghetti_plot}
\alias{.create_spaghetti_plot}
\title{Create Spaghetti Plot}
\usage{
.create_spaghetti_plot(
  data,
  outcome,
  time_var,
  id_var,
  time_range,
  colors,
  alpha
)
}
\description{
Create Spaghetti Plot
}
\keyword{internal}
