% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_class_proportions}
\alias{get_class_proportions}
\title{Get Class Proportions from Model}
\usage{
get_class_proportions(model)
}
\arguments{
\item{model}{An lctm_model or hlme object.}
}
\value{
A named numeric vector of class proportions.
}
\description{
Calculates the proportion of observations in each class.
}
\keyword{internal}
