% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_lctm_cleaned}
\alias{new_lctm_cleaned}
\title{Create a new lctm_cleaned object}
\usage{
new_lctm_cleaned(
  data,
  outcome,
  time_var,
  id_var,
  sex_var = NULL,
  n_removed = 0L,
  subjects_removed = 0L
)
}
\arguments{
\item{data}{Cleaned data frame}

\item{outcome}{Name of outcome variable}

\item{time_var}{Name of time variable}

\item{id_var}{Name of ID variable}

\item{sex_var}{Name of sex variable (optional)}

\item{n_removed}{Number of rows removed during cleaning}

\item{subjects_removed}{Number of subjects removed}
}
\value{
An lctm_cleaned object
}
\description{
Create a new lctm_cleaned object
}
\keyword{internal}
