% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_lctm_initial}
\alias{new_lctm_initial}
\title{Create a new lctm_initial object}
\usage{
new_lctm_initial(
  data,
  outcome,
  time_var,
  id_var,
  sex_var = NULL,
  degree = 2L,
  k = 2L,
  knots = NULL,
  initial_model = NULL,
  base_model = NULL,
  plots = list()
)
}
\arguments{
\item{data}{Data frame used for fitting}

\item{outcome}{Name of outcome variable}

\item{time_var}{Name of time variable}

\item{id_var}{Name of ID variable}

\item{sex_var}{Name of sex variable (optional)}

\item{degree}{Polynomial degree}

\item{k}{Number of classes in initial model}

\item{initial_model}{The hlme object (fitted with random = ~1)}

\item{base_model}{The ng=1 model for starting values}

\item{plots}{Named list of ggplot objects}
}
\value{
An lctm_initial object
}
\description{
Create a new lctm_initial object
}
\keyword{internal}
