% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_lctm_model}
\alias{new_lctm_model}
\title{Create a new lctm_model object}
\usage{
new_lctm_model(
  model,
  k,
  model_type,
  bic,
  class_proportions,
  setup,
  use_splines = FALSE,
  linear = FALSE
)
}
\arguments{
\item{model}{The lcmm hlme model object}

\item{k}{Number of classes}

\item{model_type}{Model type ("E" or "F")}

\item{bic}{BIC value}

\item{class_proportions}{Proportion in each class}

\item{setup}{The lctm_setup object used}

\item{use_splines}{Whether splines were used}

\item{linear}{Whether linear (vs quadratic) was used}
}
\value{
An lctm_model object
}
\description{
Create a new lctm_model object
}
\keyword{internal}
