% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_lctm_result}
\alias{new_lctm_result}
\title{Create a new lctm_result object}
\usage{
new_lctm_result(
  best_model,
  best_k,
  best_model_type,
  adequacy,
  bic_table,
  all_models = NULL,
  class_assignments = NULL,
  search_history = NULL
)
}
\arguments{
\item{best_model}{The best lctm_model object}

\item{best_k}{Optimal number of classes}

\item{best_model_type}{Which model type ("A" or "B")}

\item{adequacy}{lctm_adequacy object for best model}

\item{bic_table}{Data frame with all BIC comparisons}

\item{all_models}{List of all fitted lctm_model objects}

\item{class_assignments}{Data frame with individual class assignments}

\item{search_history}{Data frame documenting the search process}
}
\value{
An lctm_result object
}
\description{
Create a new lctm_result object
}
\keyword{internal}
