% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_lctm_setup}
\alias{new_lctm_setup}
\title{Create a new lctm_setup object}
\usage{
new_lctm_setup(
  data,
  outcome,
  time_var,
  id_var,
  degree = 2L,
  bic_table = NULL,
  k_ranking = NULL,
  base_model = NULL
)
}
\arguments{
\item{data}{Validated data frame}

\item{outcome}{Name of outcome variable}

\item{time_var}{Name of time variable}

\item{id_var}{Name of ID variable}

\item{degree}{Polynomial degree (1 = linear, 2 = quadratic)}

\item{bic_table}{Data frame with BIC comparisons}

\item{k_ranking}{Vector of K values ranked by BIC}

\item{base_model}{The base model (ng=1) used for initial values}
}
\value{
An lctm_setup object
}
\description{
Create a new lctm_setup object
}
\keyword{internal}
