% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_lctm_adequacy}
\alias{validate_lctm_adequacy}
\title{Validate an lctm_adequacy object}
\usage{
validate_lctm_adequacy(x)
}
\arguments{
\item{x}{An lctm_adequacy object to validate}
}
\value{
The validated object (invisibly) or throws an error
}
\description{
Validate an lctm_adequacy object
}
\keyword{internal}
