% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_lctm_data}
\alias{validate_lctm_data}
\title{Validate Input Data for LCTM}
\usage{
validate_lctm_data(data, outcome, time_var, id_var)
}
\arguments{
\item{data}{Data frame to validate}

\item{outcome}{Name of outcome variable}

\item{time_var}{Name of time variable}

\item{id_var}{Name of ID variable}
}
\value{
The validated data frame (possibly with minor cleaning)
}
\description{
Checks that input data meets requirements for LCTM analysis.
}
\keyword{internal}
