% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_lctm_model}
\alias{validate_lctm_model}
\title{Validate an lctm_model object}
\usage{
validate_lctm_model(x)
}
\arguments{
\item{x}{An lctm_model object to validate}
}
\value{
The validated object (invisibly) or throws an error
}
\description{
Validate an lctm_model object
}
\keyword{internal}
