% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_lctm_result}
\alias{validate_lctm_result}
\title{Validate an lctm_result object}
\usage{
validate_lctm_result(x)
}
\arguments{
\item{x}{An lctm_result object to validate}
}
\value{
The validated object (invisibly) or throws an error
}
\description{
Validate an lctm_result object
}
\keyword{internal}
