% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_lctm_setup}
\alias{validate_lctm_setup}
\title{Validate an lctm_setup object}
\usage{
validate_lctm_setup(x)
}
\arguments{
\item{x}{An lctm_setup object to validate}
}
\value{
The validated object (invisibly) or throws an error
}
\description{
Validate an lctm_setup object
}
\keyword{internal}
